refactor(schema): consolidate converter files into converters/
Moves json-schema and arktype-fallback converters into the converters/
subfolder for consistent organization. All schema-to-X transformations
now live in one place.
