+++ packages/epicenter/src/cli/standard-json-schema-to-yargs.ts
@@ -1,7 +1,7 @@
 import type { StandardJSONSchemaV1 } from '../core/schema/standard-schema';
 import type { JsonSchema } from 'arktype';
 import type { Argv } from 'yargs';
-import { generateJsonSchema } from '../core/schema/generate-json-schema';
+import { generateJsonSchema } from '../core/schema';
 
 // =============================================================================
 // Type Guards for JsonSchema discriminated union
